requirements.txt
streamlit
langchain_ollama
langchain_core
langchain_community
